fix default option modifyed issue
