+++ lib/run.js
@@ -3,32 +3,15 @@
 const EE = require('events').EventEmitter
 const URL = require('url')
 const hdr = require('hdr-histogram-js')
+const clone = require('clone')
 const timestring = require('timestring')
 const Client = require('./httpClient')
+const DefaultOptions = require('./defaultOptions')
 const xtend = require('xtend')
 const histUtil = require('hdr-histogram-percentiles-obj')
 const reInterval = require('reinterval')
 const histAsObj = histUtil.histAsObj
 const addPercentiles = histUtil.addPercentiles
-
-const defaultOptions = {
-  headers: {},
-  body: Buffer.alloc(0),
-  method: 'GET',
-  duration: 10,
-  connections: 10,
-  pipelining: 1,
-  timeout: 10,
-  maxConnectionRequests: 0,
-  maxOverallRequests: 0,
-  connectionRate: 0,
-  overallRate: 0,
-  amount: 0,
-  reconnectRate: 0,
-  forever: false,
-  idReplacement: false,
-  requests: [{}]
-}
 
 function run (opts, cb) {
   const cbPassedIn = (typeof cb === 'function')
@@ -69,7 +52,7 @@ function run (opts, cb) {
     0  // 5xx
   ]
 
-  opts = xtend(defaultOptions, opts)
+  opts = xtend(clone(DefaultOptions), opts)
 
   // do error checking, if error, return
   if (checkOptsForErrors()) return
